engine upgrade

+++ Engine/src/buttons/ButtonText.h
@@ -41,8 +41,8 @@ public:
 
 		Texture textTex = textTools().getTextField(text, &font, ColorA(1,0,0,1));
 		gl::pushMatrices();			
-			float shiftX = (buttonArea.getWidth() - textTex.getWidth()) * 0.5;
-			float shiftY = (buttonArea.getHeight() - textTex.getHeight()) * 0.5;
+			double shiftX = (buttonArea.getWidth() - textTex.getWidth()) * 0.5;
+			double shiftY = (buttonArea.getHeight() - textTex.getHeight()) * 0.5;
 			gl::translate(shift);
 			gl::translate(shiftX, shiftY);
 			gl::draw(textTex);
